1, use _countof instead of sizeof/sizeof 2, add string length attribute to pr file 3, modify prgen to fit for 2 4, now client can connect to gate server
vet/app: me

+++ trunk/ServerBase/common_config.cpp
@@ -68,7 +68,7 @@ bool CommonConfig::LoadConfig()
 		item.m_strServerName[iRet] = _T('\0');
 		
 		// server id
-		item.m_iServerId = (uint16)atoi(pServerElement->Attribute("id"));
+		item.m_iServerId = (uint8)atoi(pServerElement->Attribute("id"));
 
 		// peer ip
 		item.m_iPeerIP = inet_addr(pServerElement->Attribute("PeerIP"));
